Fixed the naming of components and improved UI design
