--- src/app/components/farmer/crop-info-form/disabledform-crop-info.component.css
@@ -1,56 +1,0 @@
-.container {
-  
-    border-color: var(--light-grey);
-    border: 1px solid #ccc;
-    background-color: var(--light-3);
-  }
-  #container{
-    background-color: var(--light-1);
-  }
-  .form-control {
-    border-color: var(--lighter-grey);
-  }
-  
-  
-  .button-container{
-    display: flex;
-    justify-content: center;
-    
-  }
-  .btn{
-    margin: 30px;
-  }
- hr {
-   
-    height: 2px;
-    border-color: var(--grey);
-    border: 1px solid #ccc;
-    margin-bottom: 19px;
-    }
-
-    .profile-heading{
-        display: flex;
-        align-items: center;
-        width:100%;
-      }
-
-      .heading{
-        display:flex;
-        align-items: center;
-        justify-content: flex-start;
-        width: 100%;
-      }
-      .action{
-        display:flex;
-        align-items: center;
-        justify-content: flex-end;
-        width: 100%;
-      }
-       /* CSS Styles for Success and Error States */
-.has-success .form-control {
-  border-color: green;
-}
-
-.has-error .form-control {
-  border-color: red;
-}
