+++ src/app/components/farmer/edit-all-form/edit-all-form.component.css
@@ -1,0 +1,8 @@
+/* --------------------------------
+      Created by Nkadimeng Kamogelo
+    ---------------------------------*/
+
+/* Styling for FontAwesome icons */
+fa-icon {
+    color: var(--grey);
+}
