+++ src/app/components/farmer/edit-all-form/edit-all-form.component.html
@@ -1,3 +1,5 @@
+
+
 <div class="edit-icon-container">
     <button class="btn icon-btn " type="button">
       <fa-icon icon="pen" (click)="onEditClicked()"></fa-icon>
