+++ src/app/components/farmer/edit-all-form/edit-all-form.component.ts
@@ -1,26 +1,34 @@
+ /* --------------------------------
+      Created by Nkadimeng Kamogelo
+    ---------------------------------*/
 import { Component, EventEmitter, Output } from '@angular/core';
 import { FormGroup } from '@angular/forms';
 
 @Component({
   selector: 'app-enableform-crop-info',
-  templateUrl: './enableform-crop-info.component.html',
-  styleUrls: ['./enableform-crop-info.component.css']
+  templateUrl: './edit-all-form.component.html',
+  styleUrls: ['./edit-all-form.component.css']
 })
 export class EnableformCropInfoComponent {
+
+  // Declare Output properties to emit events
   @Output() enableFields = new EventEmitter<void>();
   @Output() saveData = new EventEmitter<any>();
   @Output() cancelData = new EventEmitter<any>();
   myForm!: FormGroup;
 
+
+   // Emit the enableFields event when "Edit" is clicked
   onEditClicked() {
     this.enableFields.emit();
   }
 
-
+  // Emit the saveData event with myForm value when "Save" is clicked
   onSaveClicked() {
     this.saveData.emit(this.myForm.value);
   }
-
+  
+  // Emit the cancelData event with myForm value when "Cancel" is clicked
   onCancelClicked() {
     this.cancelData.emit(this.myForm.value);
   }
