--- src/app/components/farmer/enableform-crop-info/enableform-crop-info.component.css
@@ -1,4 +1,0 @@
-fa-icon{
-    color: var(--grey);
-    
-}
