+++ src/app/components/farmer/equipment-create/equipment-create.component.html
@@ -1,3 +1,4 @@
+
 <div class="page-body justify-center">
     <!-- <div class="col-md-6 col-sm-12 mobile-hidden">
         <img src="assets/images/background/barn.webp" alt="Avatar">
