+++ src/app/components/farmer/equipment-info-form/equipment-info-form.component.html
@@ -1,3 +1,5 @@
+
+
 <div class="container">
   <div class="profile-heading">
     <div class="heading">
