+++ src/app/components/farmer/equipment-info-form/equipment-info-form.component.ts
@@ -10,28 +10,7 @@ import { FormGroup } from '@angular/forms';
   styleUrls: ['./equipment-info-form.component.css']
 })
 export class EquipmentInfoComponent {
-  counter(i: number) {
-    return new Array(i);
-}
-  isDisabled!: boolean;
-  myForm!: FormGroup;
-  farmerData: any;
-  editedData: any;
 
-  enableFields() {
-    this.isDisabled = false; // Enable the fields by setting isDisabled to false
-    this.myForm.enable(); // Enable the formGroup
-  }
-  saveFields() {
-    this.editedData = this.myForm.value; 
-    this.isDisabled = true; 
-    }
-
-    onSaveClicked(formData: any) {
-      this.farmerData = formData;
-      this.isDisabled = true;
-      this.myForm.disable();
-    }
 
 
 }
