+++ src/app/components/farmer/equipment-table/equipment-table.component.html
@@ -1,5 +1,6 @@
 
 
+
 <div class="container record-table">
     <div class="add-icon-container">
         <button class="btn icon-btn btn" type="button" [routerLink]="['/equipment-create']">
