--- src/app/components/farmer/farm-info-form/disabledform-farm-info.component.css
@@ -1,68 +1,0 @@
-
-
-.container {
-  
-    border-color: var(--light-grey);
-    border: 1px solid #ccc;
-    background-color: var(--light-3);
-  }
-  #container{
-    background-color: var(--light-1);
-  }
-  .form-control {
-    border-color: var(--lighter-grey);
-  }
-  
-  hr {
-     
-    height: 2px;
-    border-color: var(--grey);
-    border: 1px solid #ccc;
-    margin-bottom: 20px;
-    }
-  
-    .button-container{
-      display: flex;
-      justify-content: center;
-      
-    }
-    .btn{
-      margin: 30px;
-    }
-  
-  .file-container input[type="file"] {
-      margin-top: 5px;
-    }
-    
-  
-  .profile-heading{
-    display: flex;
-    align-items: center;
-    width:100%;
-  }
-  
-  .heading{
-    display:flex;
-    align-items: center;
-    justify-content: flex-start;
-    width: 100%;
-  }
-  
-  
-  .action{
-    display:flex;
-    align-items: center;
-    justify-content: flex-end;
-    width: 100%;
-  }
-  /* CSS Styles for Success and Error States */
-.has-success .form-control {
-  border-color: green;
-}
-
-.has-error .form-control {
-  border-color: red;
-}
-  
-  
-
