+++ src/app/components/farmer/farm-info-form/farm-info-form.component.html
@@ -1,29 +1,3 @@
-<!-- <div class="outer">
-<div class="container">
-    <form class="form-input" [formGroup]="myForm" disabled>
-      <div class="form-group">
-        <label>First Name:</label>
-        <input type="text" class="form-control" formControlName="firstName">
-      </div>
-      <div class="form-group">
-        <label>Last Name:</label>
-        <input type="text" formControlName="lastName">
-      </div>
-      <div class="form-group">
-        <label>Email:</label>
-        <input type="email" formControlName="email">
-      </div>
-      <div class="form-group">
-        <label>Phone Number:</label>
-        <input type="text" formControlName="phoneNumber">
-      </div>
-    </form>
-  </div>
-</div>
-   -->
-   <!-- personal-info-form.component.html -->
-<!-- disabledform-personal-info.component.html -->
-
 
 
 <div class="container" id="container">
