--- src/app/components/farmer/personal-info-form/feature-disabledform-personal-info.component.css
@@ -1,86 +1,0 @@
-
-
-.container {
-  
-  border-color: var(--light-grey);
-  border: 1px solid #ccc;
-  background-color: var(--light-3);
-}
-#container{
-  background-color: var(--light-1);
-}
-
-.form-control {
-  border-color: var(--lighter-grey);
-}
-
-hr {
-   
-  height: 2px;
-  border-color: var(--grey);
-  border: 1px solid #ccc;
-  margin-bottom: var(--medium-s);
-  }
-
-
-
-  .file-container {
-    border: 1px solid #ccc; /* Add grey borders */
-    padding: 10px; /* Add padding to create some space around the content */
-    width: 100%; /* Set the desired width for the container */
-    background-color: #f9f9f9; 
-    border-radius: 3px;
-    
-  
-  }
-  .button-container{
-    display: flex;
-    justify-content: center;
-    
-  }
-  .btn{
-    margin: 30px;
-  }
-  #document{
-    width: 75%;
-    height: 10%;
-  }
-  .file-container label {
-    font-weight: bold; /* Make the label text bold */
-  }
-  
-  .file-container input[type="file"] {
-    margin-top: 5px;
-  }
-
-
-.profile-heading{
-  display: flex;
-  align-items: center;
-  width:100%;
-}
-
-.heading{
-  display:flex;
-  align-items: center;
-  justify-content: flex-start;
-  width: 100%;
-}
-
-.action{
-  display:flex;
-  align-items: center;
-  justify-content: flex-end;
-  width: 100%;
-}
-@media (max-width: 767px) {
-  
-}
-/* CSS Styles for Success and Error States */
-.has-success .form-control {
-  border-color: green;
-}
-
-.has-error .form-control {
-  border-color: red;
-}
