+++ src/app/pages/portfolio-view-info/portfolio-view-info.component.ts
@@ -1,3 +1,7 @@
+
+/* --------------------------------
+      Created by Nkadimeng Kamogelo
+    ---------------------------------*/
 import { Component } from '@angular/core';
 import { FormBuilder } from '@angular/forms';
