+++ src/app/services/portfolio/portfolio-service.service.ts
@@ -1,33 +1,32 @@
 import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
-import { ApiService } from '../api/api.service';
-import { Asset } from 'src/app/models/asset';
-
+import { ApiService } from '../api/api.service'; // Importing ApiService
+import { Asset } from 'src/app/models/asset'; // Importing Asset model
 
 @Injectable({
   providedIn: 'root'
 })
 export class PortfolioServiceService {
   
-  private assets: Asset[] = [];
+  private assets: Asset[] = []; // Private property to store asset data
   constructor(private _http: HttpClient, private _apiService: ApiService) { 
-    this._apiService.getAllEquipment().subscribe((data: any) => {
-     this.assets = data;
-    });
+    // Constructor initializes the service with HttpClient and ApiService
 
+    // Fetch equipment data from the API using ApiService
+    this._apiService.getAllEquipment().subscribe((data: any) => {
+      this.assets = data; // Assign fetched data to the 'assets' property
+    });
   }
 
+  // Method to generate an ID based on the length of 'assets' array
   generateId() {
     const id: number = this.assets.length;
-
     return id;
-}
-generateFarmId() {
+  }
 
-  const farm_id: number = this.assets.length;
-  return farm_id;
-}
-
-
- 
+  // Method to generate a farm ID based on the length of 'assets' array
+  generateFarmId() {
+    const farm_id: number = this.assets.length;
+    return farm_id;
+  }
 }
