+++ src/app/services/portfolio/progress.service.ts
@@ -7,14 +7,17 @@ import { BehaviorSubject } from 'rxjs';
 export class ProgressService {
     constructor() {}
 
+    // BehaviorSubject instances to track completion status
     private personalInfoCompleted = new BehaviorSubject<boolean>(false);
     private cropInfoCompleted = new BehaviorSubject<boolean>(false);
     private farmInfoCompleted = new BehaviorSubject<boolean>(false);
 
+    // Observable streams for subscribers
     personalInfoCompleted$ = this.personalInfoCompleted.asObservable();
     cropInfoCompleted$ = this.cropInfoCompleted.asObservable();
-    farmInfoCompleted$ = this.farmInfoCompleted.asObservable(); //
+    farmInfoCompleted$ = this.farmInfoCompleted.asObservable();
 
+    // Methods to set completion status and notify subscribers
     setPersonalInfoCompleted(status: boolean) {
         this.personalInfoCompleted.next(status);
     }
@@ -22,6 +25,7 @@ export class ProgressService {
     setCropInfoCompleted(status: boolean) {
         this.cropInfoCompleted.next(status);
     }
+
     setFarmInfoCompleted(status: boolean) {
         this.farmInfoCompleted.next(status);
     }
